Add core/: the ports and the vocabulary they speak
Introduce a module that everything else depends on and that depends on
nothing, and move the shared vocabulary into it. This is preparation only —
no behavior changes.

  core/types.rs    moved from src/types.rs, plus a typed Message enum for
                   the messages the server emits (unblock previously had no
                   type and was built with a json! literal)
  core/address.rs  moved out of transport/mod.rs verbatim. The server and
                   the reference model both call is_valid_address, so eight
                   call sites in server.rs and oracle.rs were reaching into
                   an adapter to validate a string.
  core/server.rs   trait ResonateServer
  core/worker.rs   trait ResonateWorker
  core/router.rs   trait ResonateRouter
  core/mod.rs      struct Unavailable — the one out-of-band failure, shared
                   by all three ports and deliberately not in types.rs
                   because it never crosses the wire.

Nothing implements the ports yet, hence the module-level allow(dead_code);
it comes off as the adapters land.
