+++ src/lib.rs
@@ -1,10 +1,10 @@
 pub mod auth;
 pub mod config;
+pub mod core;
 pub mod metrics;
 pub mod oracle;
 pub mod persistence;
 pub mod processing;
 pub mod server;
 pub mod transport;
-pub mod types;
 pub mod util;
