+++ src/persistence/persistence_postgres.rs
@@ -6,7 +6,7 @@ use super::{
     TaskFenceSettleParams, TaskFulfillParams, TaskFulfillResult, TaskHaltResult, TaskReleaseResult,
     TaskSuspendResult,
 };
-use crate::types::{
+use crate::core::types::{
     PromiseRecord, PromiseState, PromiseValue, ScheduleRecord, Snapshot, SnapshotCallback,
     SnapshotListener, SnapshotMessage, SnapshotPromiseTimeout, SnapshotTaskTimeout, TaskRecord,
     TaskState,
